+++ NEWS.md
@@ -1,6 +1,13 @@
 # powdR 1.3.0.999
 
 ## New features
+* The `afps()` function now accepts a logical `skip_nnls` argument. When
+  `skip_nnls = TRUE`, the non-negative least squares part of phase selection
+  is skipped.
+
+# powdR 1.3.0
+
+## New features
 * Vignettes updated to account for all new features detailed below.
 * Twenty one additional reference patterns associated for soils from the
   Africa Soil Information Service XRPD database now added as `afsis` data.